Restructured file structure for better organization
