+++ CMakeLists.txt
@@ -1,55 +1,69 @@
-cmake_minimum_required(VERSION 3.0.0)
+cmake_minimum_required(VERSION 3.13)
 set(CMAKE_TOOLCHAIN_FILE "C:\\vcpkg\\vcpkg\\scripts\\buildsystems\\vcpkg.cmake" )
-project(Cadera VERSION 0.0.1)
+project(Cadera VERSION 0.1.0)
 
 include(CTest)
 enable_testing()
 
-
-
-add_executable(Cadera   src/Cadera.cpp
-                        src/CADRender.cpp  
-                        src/TextRender.cpp  
-                        src/RenderUtil.cpp  
-                        src/Selection.cpp
-                        src/Camera.cpp   
-                        src/Callbacks.cpp
-                        src/ux/gui.cpp
-                        src/Sketch.cpp
-                        src/Sketcher.cpp
-                        src/SketchSolver.cpp
-                        src/Point.cpp
-                        src/Model.cpp
-                        src/Feature.cpp
-                        src/Relation.cpp
-                        src/grid.cpp      
-                        src/Main.cpp
-                       
-                      
+#=========================================
+# Targets
+#=========================================
+add_executable(CaderaApp    src/Cadera.cpp
+                            src/Selection.cpp
+                            src/Callbacks.cpp
+                            src/Model.cpp    
+                            src/Main.cpp
+                                     
 )
-    
 
-add_executable(tester test/tester.cpp)
-add_test(tester tester)
 
-target_precompile_headers(Cadera PRIVATE include/pch.hpp)
-target_include_directories(Cadera PRIVATE include/)
+add_subdirectory(src/render)
+add_subdirectory(src/sketch)
+add_subdirectory(src/ux)
 
+set_target_properties(CaderaApp PROPERTIES
+    OUTPUT_NAME CaderaApp
+    EXPORT_NAME CaderaApp
+)
+
+target_compile_definitions(CaderaApp PUBLIC)
+
+#=========================================
+# Precompiled Headers / Include Directories
+#=========================================
+target_precompile_headers(CaderaApp PRIVATE src/pch.hpp)
+
+target_include_directories(CaderaApp PRIVATE src/)
+
+#=========================================
+# Libraries
+#=========================================
 find_package(Vulkan REQUIRED)
-target_link_libraries(Cadera PRIVATE Vulkan::Vulkan)
+target_link_libraries(CaderaApp PRIVATE Vulkan::Vulkan)
 
 find_package(glfw3 CONFIG REQUIRED)
-target_link_libraries(Cadera PRIVATE glfw)
+target_link_libraries(CaderaApp PRIVATE glfw)
 
 find_package(Stb REQUIRED)
-target_include_directories(Cadera PRIVATE ${Stb_INCLUDE_DIR})
+target_include_directories(CaderaApp PRIVATE ${Stb_INCLUDE_DIR})
 
 find_package(Microsoft.GSL CONFIG REQUIRED)
-target_link_libraries(Cadera PRIVATE Microsoft.GSL::GSL)
+target_link_libraries(CaderaApp PRIVATE Microsoft.GSL::GSL)
 
 find_package(imgui CONFIG REQUIRED)
-target_link_libraries(Cadera PRIVATE imgui::imgui)
+target_link_libraries(CaderaApp PRIVATE imgui::imgui)
 
+
+#=========================================
+# Testing
+#=========================================
+include_directories(test)
+
+
+
+#=========================================
+# Packaging
+#=========================================
 set(CPACK_PROJECT_NAME ${PROJECT_NAME})
 set(CPACK_PROJECT_VERSION ${PROJECT_VERSION})
 include(CPack)
