+++ src/Selection.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "Sketch.hpp"
+#include "sketch/SketchSolver.hpp"
 
 
 namespace CADERA_APP_NAMESPACE {
