+++ src/render/CADRender.hpp
@@ -2,7 +2,7 @@
 
 #include "callbacks.hpp"
 #include "Camera.hpp"
-#include "SketchSolver.hpp"
+
