+++ src/render/CMakeLists.txt
@@ -1,0 +1,11 @@
+cmake_minimum_required(VERSION 3.13)
+
+
+
+target_sources(CaderaApp PRIVATE ${CMAKE_CURRENT_LIST_DIR}/CADRender.cpp
+                                 ${CMAKE_CURRENT_LIST_DIR}/Camera.cpp
+                                 ${CMAKE_CURRENT_LIST_DIR}/RenderUtil.cpp
+                                 ${CMAKE_CURRENT_LIST_DIR}/TextRender.cpp
+                                 ${CMAKE_CURRENT_LIST_DIR}/Grid.cpp
+                                 )
+
