+++ src/render/TextRender.cpp
