+++ test/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.13)
+
+
+add_executable(tester test/tester.cpp)
+add_test(tester tester)
+
